+++ packages/opencensus-instrumentation-http/src/types.ts
@@ -14,12 +14,19 @@
  * limitations under the License.
  */
 
-import {ClientRequest, IncomingMessage} from 'http';
+import {CustomAttributeFunction, Span} from '@opencensus/core';
+import {ClientRequest, IncomingMessage, ServerResponse} from 'http';
 
 export type IgnoreMatcher<T> =
     string|RegExp|((url: string, request: T) => boolean);
 
+export interface HttpCustomAttributeFunction extends CustomAttributeFunction {
+  (span: Span, request: ClientRequest|IncomingMessage,
+   response: IncomingMessage|ServerResponse): void;
+}
+
 export type HttpPluginConfig = {
   ignoreIncomingPaths?: Array<IgnoreMatcher<IncomingMessage>>;
   ignoreOutgoingUrls?: Array<IgnoreMatcher<ClientRequest>>;
+  applyCustomAttributesOnSpan?: HttpCustomAttributeFunction;
 };